Revert compatibility break with NodePath and ConfigurationTransformation
